metasejong-chatbot/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pymysql
httpx
langchain
langchain-openai
faiss-cpu
tiktoken